+++ web/static/js/brando/brando/components/accordion.js
@@ -6,9 +6,16 @@ class Accordion {
   static setup() {
     var that = this;
     $(document).ready(function () {
+      var hash = document.location.hash;
       $('.accordion-tabs-minimal').each(function() {
-        if (!document.location.hash) {
-          $(this).children('li').first().children('a').addClass('is-active').next().addClass('is-open').show();
+        if (!hash) {
+          let $link = $(this).children('li').first().children('a');
+          let $linkSibling = $link.next();
+          $link.addClass('is-active');
+          $linkSibling.addClass('is-open').show();
+        } else {
+          let $link = $('#tab-' + hash.replace('#', ''));
+          that.activateTab($link);
         }
       });
 
@@ -20,16 +27,20 @@ class Accordion {
   }
 
   static activateTab(obj) {
-    if (!$(obj).hasClass('is-active')) {
-      // remove `tab-` from obj id
-      document.location.hash = $(obj).attr('id').replace('tab-', '');
-      var accordionTabs = $(obj).closest('.accordion-tabs-minimal');
-      accordionTabs.find('.is-open').removeClass('is-open').hide();
+    let $obj = $(obj),
+        $accordionTabs = $obj.closest('.accordion-tabs-minimal'),
+        $openTabs = $accordionTabs.find('.is-open'),
+        $activeTabs = $accordionTabs.find('.is-active'),
+        $tabContent = $obj.next();
 
-      $(obj).next().toggleClass('is-open').toggle();
-      accordionTabs.find('.is-active').removeClass('is-active');
-      $(obj).addClass('is-active');
+    if (!$obj.hasClass('is-active')) {
+      document.location.hash = $obj.attr('id').replace('tab-', '');
     }
+
+    $openTabs.removeClass('is-open').hide();
+    $tabContent.toggleClass('is-open').toggle();
+    $activeTabs.removeClass('is-active');
+    $obj.addClass('is-active');
   }
 }